Move AggregateError to primordials
