+++ lib/ours/util.js
@@ -5,7 +5,7 @@ const { format, inspect } = require('./util/inspect')
 const {
   codes: { ERR_INVALID_ARG_TYPE }
 } = require('./errors')
-const { kResistStopPropagation, SymbolDispose } = require('./primordials')
+const { kResistStopPropagation, AggregateError, SymbolDispose } = require('./primordials')
 const AbortSignal = globalThis.AbortSignal || require('abort-controller').AbortSignal
 const AbortController = globalThis.AbortController || require('abort-controller').AbortController
 const AsyncFunction = Object.getPrototypeOf(async function () {}).constructor
@@ -30,22 +30,6 @@ const validateAbortSignal = (signal, name) => {
 const validateFunction = (value, name) => {
   if (typeof value !== 'function') {
     throw new ERR_INVALID_ARG_TYPE(name, 'Function', value)
-  }
-}
-
-// This is a simplified version of AggregateError
-class AggregateError extends Error {
-  constructor(errors) {
-    if (!Array.isArray(errors)) {
-      throw new TypeError(`Expected input to be an Array, got ${typeof errors}`)
-    }
-    let message = ''
-    for (let i = 0; i < errors.length; i++) {
-      message += `    ${errors[i].stack}\n`
-    }
-    super(message)
-    this.name = 'AggregateError'
-    this.errors = errors
   }
 }
 module.exports = {
