+++ src/primordials.js
@@ -8,7 +8,26 @@
   Don't try to replace with the original file and keep it up to date with the upstream file.
 */
 
+// This is a simplified version of AggregateError
+class AggregateError extends Error {
+  constructor(errors) {
+    if (!Array.isArray(errors)) {
+      throw new TypeError(`Expected input to be an Array, got ${typeof errors}`)
+    }
+
+    let message = ''
+    for (let i = 0; i < errors.length; i++) {
+      message += `    ${errors[i].stack}\n`
+    }
+
+    super(message)
+    this.name = 'AggregateError'
+    this.errors = errors
+  }
+}
+
 module.exports = {
+  AggregateError,
   ArrayIsArray(self) {
     return Array.isArray(self)
   },
